+++ src/main/java/ru/baldursgate3/tgbot/storeservice/StoreServiceApplication.java
@@ -12,3 +12,4 @@ public class StoreServiceApplication {
 	}
 
 }
+
