+++ src/main/java/ru/baldursgate3/tgbot/storeservice/controllers/UserController.java
@@ -1,4 +1,4 @@
-package ru.baldursgate3.tgbot.storeservice.jpa.controllers;
+package ru.baldursgate3.tgbot.storeservice.controllers;
 
 import org.springframework.web.bind.annotation.RestController;
