+++ src/main/java/ru/baldursgate3/tgbot/storeservice/repositories/UserRepository.java
@@ -1,7 +1,7 @@
-package ru.baldursgate3.tgbot.storeservice.jpa.repositories;
+package ru.baldursgate3.tgbot.storeservice.repositories;
 
 import org.springframework.data.jpa.repository.JpaRepository;
-import ru.baldursgate3.tgbot.storeservice.jpa.entities.User;
+import ru.baldursgate3.tgbot.storeservice.entities.User;
 
 public interface UserRepository extends JpaRepository<User, Long> {
